+++ R/openairmaps-package.R
@@ -39,7 +39,8 @@
 "_PACKAGE"
 
 ## usethis namespace: start
-#' @importFrom tibble tibble
+#' @importFrom lifecycle deprecated
 #' @importFrom rlang .data
+#' @importFrom tibble tibble
 ## usethis namespace: end
 NULL