+++ R/polar_annulusMap.R
@@ -45,13 +45,15 @@ annulusMap <- function(data,
                        collapse.control = FALSE,
                        d.icon = 200,
                        d.fig = 3.5,
-                       type = NULL,
+                       type = deprecated(),
                        ...) {
-  if (!is.null(type)) {
-    cli::cli_warn(c(
-      "!" = "{.code type} is deprecated. Different sites are now automatically identified.",
-      "i" = "Please use {.code label} and/or {.code popup} to label sites."
-    ))
+  if (lifecycle::is_present(type)) {
+    lifecycle::deprecate_soft(
+      when = "0.5.0",
+      what = "openairmaps::annulusMap(type)",
+      details = c("Different sites are now automatically detected based on latitude and longitude",
+                  "Please use the `popup` argument to create popups.")
+    )
   }
 
   # assume lat/lon
@@ -87,7 +89,6 @@ annulusMap <- function(data,
   data <-
     prepMapData(
       data = data,
-      type = type,
       pollutant = pollutant,
       control = control,
       "wd",