+++ R/polar_diffMap.R
@@ -66,13 +66,15 @@ diffMap <- function(before,
                     collapse.control = FALSE,
                     d.icon = 200,
                     d.fig = 3.5,
-                    type = NULL,
+                    type = deprecated(),
                     ...) {
-  if (!is.null(type)) {
-    cli::cli_warn(c(
-      "!" = "{.code type} is deprecated. Different sites are now automatically identified.",
-      "i" = "Please use {.code label} and/or {.code popup} to label sites."
-    ))
+  if (lifecycle::is_present(type)) {
+    lifecycle::deprecate_soft(
+      when = "0.5.0",
+      what = "openairmaps::diffMap(type)",
+      details = c("Different sites are now automatically detected based on latitude and longitude",
+                  "Please use the `popup` argument to create popups.")
+    )
   }
 
   # assume lat/lon