+++ R/polar_percentileMap.R
@@ -44,13 +44,15 @@ percentileMap <- function(data,
                           collapse.control = FALSE,
                           d.icon = 200,
                           d.fig = 3.5,
-                          type = NULL,
+                          type = deprecated(),
                           ...) {
-  if (!is.null(type)) {
-    cli::cli_warn(c(
-      "!" = "{.code type} is deprecated. Different sites are now automatically identified.",
-      "i" = "Please use {.code label} and/or {.code popup} to label sites."
-    ))
+  if (lifecycle::is_present(type)) {
+    lifecycle::deprecate_soft(
+      when = "0.5.0",
+      what = "openairmaps::percentileMap(type)",
+      details = c("Different sites are now automatically detected based on latitude and longitude",
+                  "Please use the `popup` argument to create popups.")
+    )
   }
 
   # assume lat/lon
@@ -82,7 +84,6 @@ percentileMap <- function(data,
   data <-
     prepMapData(
       data = data,
-      type = type,
       pollutant = pollutant,
       control = control,
       "wd",