+++ R/polar_polarMap.R
@@ -54,8 +54,9 @@
 #'   inches. This will affect the resolution of the markers on the map.
 #'   Alternatively, a vector in the form `c(width, height)` can be provided if a
 #'   non-circular marker is desired.
-#' @param type Deprecated. Please use `label` and/or `popup` to label different
-#'   sites.
+#' @param type `r lifecycle::badge("deprecated")`. Different sites are now
+#'   automatically detected based on latitude and longitude. Please use `label`
+#'   and/or `popup` to label different sites.
 #' @inheritDotParams openair::polarPlot -mydata -pollutant -x -limits -type
 #'   -cols -key -alpha -plot
 #' @return A leaflet object.
@@ -89,13 +90,15 @@ polarMap <- function(data,
                      collapse.control = FALSE,
                      d.icon = 200,
                      d.fig = 3.5,
-                     type = NULL,
+                     type = deprecated(),
                      ...) {
-  if (!is.null(type)) {
-    cli::cli_warn(c(
-      "!" = "{.code type} is deprecated. Different sites are now automatically identified.",
-      "i" = "Please use {.code label} and/or {.code popup} to label sites."
-    ))
+  if (lifecycle::is_present(type)) {
+    lifecycle::deprecate_soft(
+      when = "0.5.0",
+      what = "openairmaps::polarMap(type)",
+      details = c("Different sites are now automatically detected based on latitude and longitude",
+      "Please use the `popup` argument to create popups.")
+    )
   }
 
   # assume lat/lon